+++ jsolex-core/src/main/java/me/champeau/a4j/jsolex/processing/expr/impl/Crop.java
@@ -249,7 +249,7 @@ public class Crop extends AbstractFunctionImpl {
         });
         img.findMetadata(ActiveRegions.class).ifPresent(activeRegions -> metadata.put(ActiveRegions.class, activeRegions.translate(-left, -top)));
         img.findMetadata(Flares.class).ifPresent(flares -> metadata.put(Flares.class, flares.translate(-left, -top)));
-        img.findMetadata(ReferenceCoords.class).ifPresent(coords -> metadata.put(ReferenceCoords.class, coords.addOffsetX(left).addOffsetY(top)));
+        img.findMetadata(ReferenceCoords.class).ifPresent(coords -> metadata.put(ReferenceCoords.class, coords.addOffset2D(left, top)));
         return metadata;
     }